Restructure packaging for instant
COAND-497
